requirements.txt
httpx
tenacity
playwright
jinja2
python-dotenv
pydantic
rich
tqdm
pyyaml